+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.8.2 (Dec 02, 2025)
+* Improved aggregation of capability outputs to reduce errors and improve reliability.
+* Upgrade `google` terraform provider.
+
 # 0.8.1 (Oct 22, 2025)
 * Upgraded external secrets apiVersion to `external-secrets.io/v1` (was `external-secrets.io/v1beta1`).